Add model provider info to 'new chat' page
